+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # Master
 
 - Feature: Select combinations #51
+- Feature: Snapshot select can use state class now #398
 - Performance: Run actions outside of zones #383
 - Fix: Dispatcher error propogation #376
 - Fix: Add patchState restriction errors #371
 - Fix: Remove error when loading same state multiple times #390
-- Fix: Force route navigation to run in zones
+- Fix: Force route navigation to run in zones #393
+- Fix: Selector function should still be usable as a function #398
 
 # 3.0.0/3.0.1 5/3/18