+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Master
 
+- Feature: Add development mode with freeze on state and actions
 - Feature: Storage engine migrations
 - Feature: Select combinations #51
 - Feature: Snapshot select can use state class now #398